refactor(bench): one definition per cross-arm constant, and drop the dead ends
Four arms carried private copies of the values that MUST be identical across
them. The mutation wobble was the dangerous one: `selectMutationIndices` picks
which leaves move and the harness asserts that choice cross-arm, but the
amplitude and phase step were four separate constants no check ever compared.
Drifting one would have moved the identical leaves by different distances and
reported it as an engine difference. It is now `wobbleOffsetAt` in
`shared/mutation.ts`, beside the selection it belongs to.

`sceneAssets.ts` takes over the raster assets on the same argument: the distinct
sprite texture (four copies), the digit glyph sheet, and the shared text font
size. The Phaser and Excalibur arms also stop transcribing the grid layout and
viewport constants and call `world.ts`'s helpers, so a change to the layout can
no longer move one arm's scene without moving every arm's.

Three dead ends removed while in here: `PerStepWork.active`, which nothing read;
`rowsIntoWebgl1`, an indirection over a single `push`; and the rendering CLI's
`filter` object, which was built empty and passed to a no-op filter on every run.
`chooseHeadlineNodeCount` becomes `chooseHeadlineCount` - it picks the physics
block's body count too.

Verified by running all four arms over `dynamic-heavy`, `text-static` and
`lifecycle-churn` on a real GPU: twelve cells, all `ok`, which the harness's
cross-arm determinism assertion would have refused had any arm's scene moved.
